+++ landingzones/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class LandingzonesConfig(AppConfig):
     name = 'landingzones'
+
+    def ready(self):
+        import landingzones.checks  # noqa